Resolved initialization.
This strategy calls `init` to bring the state to t=0+. The sim then
begins. The `init` function can, where appropriate, just call `update`,
or it can do things like take random draws and set up states. When we
will later want to load a state from a log and start the sim there,
we'll load all of the states, inputs, and outputs, and then begin the
sim without calling `init`.